Added "Day6.kt", did performance optimization and added dependency for "coroutines"
A new logic is formulated in the new file "Day6.kt" for the calculations of time and distance. Sequences in Kotlin are used for optimization in file "Day5.kt" for better performance and efficiency. In addition to this, the jetbrains.kotlinx.coroutines.core is added as a project-level dependency for effective asynchronous programming. The example text file has also been updated to day 6